+++ epam-promo/components/tables/DataTable.tsx
@@ -1,6 +1,9 @@
 import * as React from 'react';
-import { PositionValues, VirtualListRenderRowsParams, useColumnsWithFilters, IconContainer } from '@epam/uui-components';
-import { ColumnsConfig, DataRowProps, useUuiContext, uuiScrollShadows, useColumnsConfig, IEditable, DataTableState, DataTableColumnsConfigOptions, DataSourceListProps, DataColumnProps, cx, TableFiltersConfig } from '@epam/uui-core';
+import { PositionValues, VirtualListRenderRowsParams, useColumnsWithFilters,
+        IconContainer } from '@epam/uui-components';
+import { ColumnsConfig, DataRowProps, useUuiContext, uuiScrollShadows, useColumnsConfig, IEditable,
+        DataTableState, DataTableColumnsConfigOptions, DataSourceListProps, DataColumnProps,
+        cx, TableFiltersConfig, DataTableRowProps } from '@epam/uui-core';
 import { ColumnsConfigurationModal, DataTableHeaderRow, DataTableRow, DataTableMods } from './';
 import { VirtualList } from '../';
 import { ReactComponent as EmptyTableIcon } from '../../icons/empty-table.svg';
@@ -11,7 +14,7 @@ import { i18n } from "../../i18n";
 export interface DataTableProps<TItem, TId> extends IEditable<DataTableState>, DataSourceListProps, DataTableColumnsConfigOptions {
     getRows(): DataRowProps<TItem, TId>[];
     columns: DataColumnProps<TItem, TId>[];
-    renderRow?(props: DataRowProps<TItem, TId>): React.ReactNode;
+    renderRow?(props: DataTableRowProps<TItem, TId>): React.ReactNode;
     renderNoResultsBlock?(): React.ReactNode;
     onScroll?(value: PositionValues): void;
     showColumnsConfig?: boolean;
@@ -23,7 +26,7 @@ export function DataTable<TItem, TId>(props: React.PropsWithChildren<DataTablePr
     const columnsWithFilters = useColumnsWithFilters(props.columns, props.filters);
     const { columns, config, defaultConfig } = useColumnsConfig(columnsWithFilters, props.value?.columnsConfig);
 
-    const renderRow = React.useCallback((rowProps: DataRowProps<TItem, TId>) => (
+    const renderRow = React.useCallback((rowProps: DataTableRowProps<TItem, TId>) => (
         <DataTableRow
             key={ rowProps.rowKey }
             size={ props.size }