+++ uui-core/data/processing/hooks/useLazyDataSource.ts
@@ -1,6 +1,11 @@
 import { LazyDataSource, LazyDataSourceProps } from '../LazyDataSource';
 import { useMemoWithDestructor } from "../../../helpers/useMemoWithDestructor";
 
-export function useLazyDataSource<TItem, TId, TFilter>(params: LazyDataSourceProps<TItem, TId, TFilter>, deps: any[]) {
-    return useMemoWithDestructor(() => new LazyDataSource({ ...params}), (dataSource) => dataSource.destroy(), deps);
+export function useLazyDataSource<TItem, TId, TFilter>(props: LazyDataSourceProps<TItem, TId, TFilter>, deps: any[]) {
+    return useMemoWithDestructor(
+        () => new LazyDataSource({ ...props}),
+        (dataSource) => dataSource.setProps(props),
+        (dataSource) => dataSource.destroy(),
+        deps
+    );
 }